pkg/util/log: add "none" bufferedSink format option
As we introduce the ability to enable `buffering` on file-based log
sinks as experimental, there's a need for a new `BufferFormat`.

Previously, the default format was to introduce a newline between each
log message flushed to the underlying log sink as a delimiter. However,
when flushing to files, this is not desireable behavior as it causes the
log file to look like:
```
I240313 16:38:33.355448 113 1@gossip/gossip.go:1374 ⋮ [T1,Vsystem,n1] 80  node has connected to cluster via gossip

I240313 16:38:33.360943 113 kv/kvserver/stores.go:282 ⋮ [T1,Vsystem,n1] 81  wrote 0 node addresses to persistent storage
```

Empty lines do not play well with log parsers, so we introduce a new
`BufferFormat`, `BufferFmtNone`, which uses no delimiters, prefixes, or
suffixes. With this new option, log files no longer see these empty
lines.

The validation code for file log sinks forces the `BufferFmtNone` into
the buffering configuration, so this is not something that needs to
explicitly be configured by the user.

Release note: none